+++ spec/policies/user_policy_spec.rb
@@ -1,7 +1,7 @@
 require "rails_helper"
 
 RSpec.describe UserPolicy do
-  permissions :index? do
+  permissions :index?, :new?, :create? do
     it "grants access if the user has the service_support flag" do
       user = build(:user, :service_support)
       expect(described_class).to permit(user)